fast_api_student_apis/config.py
# config.py
from pydantic_settings import BaseSettings

class Settings(BaseSettings):
    db_host: str
    db_port: int
    db_name: str
    db_user: str
    db_password: str

    class Config:
        env_file = ".env"

settings = Settings()  # ✅ This line creates the instance you're trying to import