fast_api_student_apis/requirements.txt
fastapi==0.103.2
uvicorn==0.23.2
python-dotenv==1.0.0
pydantic==1.10.13
psycopg==3.1.18
pandas==2.2.2
openpyxl==3.1.2
